# specs/001-warehouse-management/contracts/openapi.yaml
openapi: 3.1.2
...
info:
  title: Warehouse Manager API
  version: 1.0.0
  description: >-
    Authoritative HTTP contract for warehouse, route, sale, reconciliation, reporting, document, user, and
    printer-profile operations. Money and quantities are decimal strings. All protected authorization and business rules
    are enforced by the API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /locations:
    get:
      tags: [ Catalog ]
      operationId: listLocations
      responses:
        '200':
          description: Stock-holding branches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationListResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List locations
      parameters:
        - name: active
          in: query
          description: Filter active or archived catalog entries; omitted returns both.
          schema:
            type: boolean
    post:
      tags: [ Catalog ]
      operationId: createLocation
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationWriteRequest'
      responses:
        '201':
          description: Location created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create location
  /categories:
    get:
      tags: [ Catalog ]
      operationId: listCategories
      responses:
        '200':
          description: Product categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List categories
      parameters:
        - name: active
          in: query
          description: Filter active or archived catalog entries; omitted returns both.
          schema:
            type: boolean
    post:
      tags: [ Catalog ]
      operationId: createCategory
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryWriteRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create category
  /units:
    get:
      tags: [ Catalog ]
      operationId: listUnits
      responses:
        '200':
          description: Product units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitListResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List units
      parameters:
        - name: active
          in: query
          description: Filter active or archived catalog entries; omitted returns both.
          schema:
            type: boolean
    post:
      tags: [ Catalog ]
      operationId: createUnit
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitWriteRequest'
      responses:
        '201':
          description: Unit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create unit
  /products:
    get:
      tags: [ Catalog ]
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: active
          in: query
          description: Filter active or archived catalog entries; omitted returns both.
          schema:
            type: boolean
      responses:
        '200':
          description: Products visible to the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List products
    post:
      tags: [ Catalog ]
      operationId: createProduct
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductWriteRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create product
  /products/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [ Catalog ]
      operationId: getProduct
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Get product
    patch:
      tags: [ Catalog ]
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated, activated, or archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update product
...
  /vehicles:
    get:
      tags: [ Catalog ]
      operationId: listVehicles
      responses:
        '200':
          description: Delivery vehicles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List vehicles
      parameters:
        - name: active
          in: query
          description: Filter active or archived catalog entries; omitted returns both.
          schema:
            type: boolean
    post:
      tags: [ Catalog ]
      operationId: createVehicle
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleWriteRequest'
      responses:
        '201':
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create vehicle
...
  /locations/{locationId}:
    parameters:
      - name: locationId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [ Catalog ]
      operationId: updateLocation
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdateRequest'
      responses:
        '200':
          description: Location updated or deactivated without deleting history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update location
  /categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [ Catalog ]
      operationId: updateCategory
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Category updated or deactivated without rewriting sale snapshots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update category
  /units/{unitId}:
    parameters:
      - name: unitId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [ Catalog ]
      operationId: updateUnit
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitUpdateRequest'
      responses:
        '200':
          description: Unit updated or deactivated without deleting history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnitResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update unit
  /vehicles/{vehicleId}:
    parameters:
      - name: vehicleId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [ Catalog ]
      operationId: updateVehicle
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUpdateRequest'
      responses:
        '200':
          description: Vehicle updated or deactivated when it has no active route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update vehicle
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-wm_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Per-session synchronizer token for an authenticated unsafe request
      schema:
        type: string
        minLength: 32
        maxLength: 512
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 120
...
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    Problem:
      description: RFC 9457 error; exact status and stable code identify the failure
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Uuid:
      type: string
      format: uuid
...
    LocationWriteRequest:
      type: object
      additionalProperties: false
      required: [ code, name ]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 120
    LocationUpdateRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion, code, name, active ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        code:
          type: string
          minLength: 1
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 120
        active:
          type: boolean
        reason:
          type: [ string, 'null' ]
          maxLength: 500
    LocationResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Location'
    LocationListResponse:
      type: object
      required: [ data ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Location'
...
    CategoryWriteRequest:
      type: object
      additionalProperties: false
      required: [ name, reportingGroup ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        reportingGroup:
          $ref: '#/components/schemas/ReportingGroup'
    CategoryUpdateRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion, name, reportingGroup, active ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 120
        reportingGroup:
          $ref: '#/components/schemas/ReportingGroup'
        active:
          type: boolean
        reason:
          type: [ string, 'null' ]
          maxLength: 500
    CategoryResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Category'
    CategoryListResponse:
      type: object
      required: [ data ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
...
    UnitWriteRequest:
      type: object
      additionalProperties: false
      required: [ code, name, quantityScale ]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 80
        quantityScale:
          type: integer
          minimum: 0
          maximum: 3
...
    UnitResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Unit'
    UnitListResponse:
      type: object
      required: [ data ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Unit'
...
    ProductWriteRequest:
      type: object
      additionalProperties: false
      required: [ sku, name, categoryId, unitId, standardUnitPrice, lowStockThreshold ]
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: [ string, 'null' ]
          maxLength: 1000
        categoryId:
          $ref: '#/components/schemas/Uuid'
        unitId:
          $ref: '#/components/schemas/Uuid'
        standardUnitPrice:
          $ref: '#/components/schemas/UnitPrice'
        lowStockThreshold:
          $ref: '#/components/schemas/Quantity'
    ProductUpdateRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion, sku, name, categoryId, unitId, standardUnitPrice, lowStockThreshold, active ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        sku:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: [ string, 'null' ]
          maxLength: 1000
        categoryId:
          $ref: '#/components/schemas/Uuid'
        unitId:
          $ref: '#/components/schemas/Uuid'
        standardUnitPrice:
          $ref: '#/components/schemas/UnitPrice'
        lowStockThreshold:
          $ref: '#/components/schemas/Quantity'
        active:
          type: boolean
        reason:
          type: [ string, 'null' ]
          maxLength: 500
    ProductResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Product'
    ProductListResponse:
      type: object
      required: [ data, page ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    VehicleWriteRequest:
      type: object
      additionalProperties: false
      required: [ code, name ]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 120
        registration:
          type: [ string, 'null' ]
          maxLength: 64
...
    VehicleResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/Vehicle'
    VehicleListResponse:
      type: object
      required: [ data ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
...